refactor(storage): move sigv4 helper to shared/
Pure stdlib signing function with no I/O — shared/'s charter, next to
image_sniff. infrastructure/storage keeps only the R2 client.
