+++ shared/sigv4.py
